Round corners
The problem here is that the paths' ends aren't oriented right (they
should be cut either vertically or horizontally, instead they're
perpendicular to the end of the last stroke). This causes ugly seams
where the path ends on itself, and also gives a bad look as all the
path's ends for the various open passages just face random ways.

By making line joins and caps *round*, we just sidestep the
directionality problem of paths' ends entirely. Everything is nice and
round and squishy and pretty.